components/back/requirements.txt
fastapi
uvicorn
requests
python-multipart
pydantic[email]
python-dotenv
